
% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add.do}
\alias{add.do}
\title{
add.do
}

\usage{
add.do(x = x, h = h, col = col, .order = NA, begin = 0, position = "fill", scale_height = 1, scale.col = NA)
}

\arguments{
  \item{x}{
  A numeric vector or a character,is circlize current sector index ,using  \code{get.current.sector.index()} to reading


}
  \item{h}{

A numeric vector or a character. If it is a character, will scale  to the height of heatmap.
}
  \item{col}{
A numeric vector or a character. If it is a character, will scale  to the height of heatmap.
Gradient

}
  \item{.order}{
A vector color or histogram and sorting of the heatmap
}
  \item{begin}{
Starting position
}
  \item{position}{
position=c("fill","stack"),
"stack" is Stacked histogram ,"fill is percentage
}
  \item{scale_height}{
Standardized height

}
  \item{scale.col}{
A color vector, name is String or number}
}

\author{
dongbin wang,chao chen
}

\examples{

# data col
data = data.frame(from = rep(letters[1:3],each=3), to = LETTERS[1:3], value = 1:3)
structure(c("red"), names = c("a"))->col


set.seed(2)
#heatmap
x=rep(c("a","b","c"),3);
col<-(c("a","b","c","e","f","a","c","c","A"));
#col= rnorm(9) ;
h=rep(c("a","b","c"),each=3);
structure(c("red","blue","grey"), names = c("a",-1,5))->scale.col
structure(c("red","blue","grey"), names = c("e","d","c"))->scale.col
.order<-c("a","b","c")

structure(c(max(col),median(col),min(col)),
          names = c("red","grey","blue"
          ))->scale.col
chord(data,col=col,scale = F)->chor

add.do(x=x,h=h,col=col,begin=-0.1,.order =.order,scale.col = scale.col ,
       scale_height = 0.5)->len


add.do(x=x,h=h,col=col,begin=0.5,.order =.order,scale.col = scale.col ,
       scale_height = 0.5)->len

#draw(h)

#draw(len)

#bar
h=c(1,2,3,0.5);
x=c("A","B","C","A");
col<-c("a","e","a","b")
structure(c("red","blue","grey"), names = c("a",-1,5))->scale.col
structure(c("red","blue","black"), names = c("e","d","c"))->scale.col
.order<-c("a","b","e")

add.do(x=x,h=h,col=col,begin=-0.1,position = "stack",
       scale_height = 0.6)->len2

add.do(x=x,h=h,col=col,begin=0.9,position = "stack",
       scale_height = 0.6)->len2

x=c("a","b","c","a");
col<-c("a","e","a","b")
structure(c("red","blue","grey"), names = c("a",-1,5))->scale.col
structure(c("red","blue","black"), names = c("e","d","c"))->scale.col
.order<-c("a","b","e")
#add.do(x=x,h=h,col=col,begin=0.5,position = "fill",scale_height = 0.4)->len3

x=c("a","b","c");h=c(1:3);col=(1:3)
structure(c("red","black","blue"), names = c(1,2.7,3))->scale.col

add.do(x=x,h=h,col=col,begin=0.6,scale_height = 0.3,
       scale.col=scale.col)->len3

add.do(x=x,h=h,col=col,begin=0.8,
       scale_height = 0.8,scale.col=scale.col)->len3




x=c("a","b","c") ;name=c("ggplot","gglen","wanf")

add.name(x,name,begin =9)



x=rep(c("a","b","c"),3);
#col= rnorm(9) ;
h=rep(c("a","b","c"),each=3);
pch<-c("a","b","c")
.order<-c("a","c","b")
structure(c("a","b","c"), names = c(1,2,3))->scale.pch


add.pch(x=x,h=h,pch=pch,begin =2,
        scale_height = 3,col="red")




}
