% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{char_one_clusterProfiler}
\alias{char_one_clusterProfiler}
\title{Title}
\usage{
char_one_clusterProfiler(
  metaProfiler,
  data = NA,
  show = 125,
  pvalue = 0.05,
  heatmap_scale_color = NA,
  bar.height.scale = 0.6,
  name.begin = 0.8,
  pathway.name = T,
  gean.name = F,
  bar.col = NA,
  scale = T,
  ...
)
}
\arguments{
\item{metaProfiler}{a metaProfiler data}

\item{data}{data is who}

\item{show}{show number}

\item{pvalue}{pvalue =0.05}

\item{heatmap_scale_color}{heatmap color}

\item{bar.height.scale}{pathways bar height}

\item{name.begin}{begin names higth}

\item{pathway.name}{if T will add pathway names}

\item{gean.name}{if T will add gean names}

\item{bar.col}{pathways bar color}

\item{scale}{scale of width}

\item{...}{}
}
\value{
list heatmap
}
\description{
Title
}
\examples{
char_one_clusterProfiler(n)\%>\%packLegend(list = .)->lgd_list_vertical 
draw(lgd_list_vertical, x = unit(4, "mm"), y = unit(4,"mm"), just = c("left", "bottom"))
}
