% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{getcol}
\alias{getcol}
\title{packagename: Package containing tools support project work.}
\usage{
getcol(col = col, scale.col = scale.col)
}
\description{
This package provides a central repository for methods to facilitate 
project work
}
