% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst.R, R/codemmu.R
\docType{class}
\name{metaProfiler-class}
\alias{metaProfiler-class}
\title{packagename: Package containing tools support project work.}
\description{
This package provides a central repository for methods to facilitate 
project work

This package provides a central repository for methods to facilitate 
project work
}
