% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_plot.R
\name{plot_add}
\alias{plot_add}
\title{packagename: Package containing tools support project work.}
\usage{
plot_add(
  f,
  begin = begin,
  scale_height = scale_height,
  scale_color = scale_color,
  metaProfiler = metaProfiler,
  group2 = group2
)
}
\description{
This package provides a central repository for methods to facilitate 
project work
}
